+++ relay-server/src/actors/processor.rs
@@ -45,7 +45,7 @@ use crate::envelope::{AttachmentType, ContentType, Envelope, Item, ItemType};
 use crate::metrics_extraction::sessions::{extract_session_metrics, SessionMetricsConfig};
 use crate::metrics_extraction::transactions::{extract_transaction_metrics, ExtractMetricsError};
 use crate::service::REGISTRY;
-use crate::statsd::{RelayCounters, RelayHistograms, RelayTimers};
+use crate::statsd::{RelayCounters, RelayTimers};
 use crate::utils::{
     self, ChunkedFormDataAggregator, EnvelopeContext, ErrorBoundary, FormDataIter, SamplingResult,
 };
@@ -346,118 +346,6 @@ fn outcome_from_parts(field: ClientReportField, reason: &str) -> Result<Outcome,
             "" => None,
             other => Some(ReasonCode::new(other)),
         })),
-    }
-}
-
-fn outcome_from_profile_error(err: relay_profiling::ProfileError) -> Outcome {
-    let discard_reason = match err {
-        relay_profiling::ProfileError::CannotSerializePayload => DiscardReason::Internal,
-        relay_profiling::ProfileError::NotEnoughSamples => DiscardReason::InvalidProfile,
-        _ => DiscardReason::ProcessProfile,
-    };
-    Outcome::Invalid(discard_reason)
-}
-
-fn track_sampling_metrics(
-    project_state: &ProjectState,
-    context: &DynamicSamplingContext,
-    event: &Event,
-) {
-    // We only collect this metric for the root transaction event, so ignore secondary projects.
-    if !project_state.is_matching_key(context.public_key) {
-        return;
-    }
-
-    let transaction_info = match event.transaction_info.value() {
-        Some(info) => info,
-        None => return,
-    };
-
-    let changes = match transaction_info.changes.value() {
-        Some(value) => value.as_slice(),
-        None => return,
-    };
-
-    let last_change = changes
-        .iter()
-        .rev()
-        // skip all broken change records
-        .filter_map(|a| a.value())
-        // skip records without a timestamp
-        .filter(|c| c.timestamp.value().is_some())
-        // take the last that did not occur when the event was sent
-        .find(|c| c.timestamp.value() != event.timestamp.value());
-
-    let source = event.get_transaction_source().as_str();
-    let platform = event.platform.as_str().unwrap_or("other");
-    let sdk_name = event.sdk_name();
-    let sdk_version = event.sdk_version();
-
-    metric!(
-        histogram(RelayHistograms::DynamicSamplingChanges) = changes.len() as u64,
-        source = source,
-        platform = platform,
-        sdk_name = sdk_name,
-        sdk_version = sdk_version,
-    );
-
-    if let Some(&total) = transaction_info.propagations.value() {
-        // If there was no change, there were no propagations that happened with a wrong name.
-        let change = last_change
-            .and_then(|c| c.propagations.value())
-            .map_or(0, |v| *v);
-
-        metric!(
-            histogram(RelayHistograms::DynamicSamplingPropagationCount) = change,
-            source = source,
-            platform = platform,
-            sdk_name = sdk_name,
-            sdk_version = sdk_version,
-        );
-
-        let percentage = match (change, total) {
-            (0, 0) => 0.0, // 0% indicates no premature changes.
-            _ => ((change as f64) / (total as f64)).min(1.0) * 100.0,
-        };
-
-        metric!(
-            histogram(RelayHistograms::DynamicSamplingPropagationPercentage) = percentage,
-            source = source,
-            platform = platform,
-            sdk_name = sdk_name,
-            sdk_version = sdk_version,
-        );
-    }
-
-    if let (Some(&start), Some(&change), Some(&end)) = (
-        event.start_timestamp.value(),
-        last_change
-            .and_then(|c| c.timestamp.value())
-            .or_else(|| event.start_timestamp.value()), // default to start if there was no change
-        event.timestamp.value(),
-    ) {
-        let delay_ms = (change - start).num_milliseconds();
-        if delay_ms >= 0 {
-            metric!(
-                histogram(RelayHistograms::DynamicSamplingChangeDuration) = delay_ms as u64,
-                source = source,
-                platform = platform,
-                sdk_name = sdk_name,
-                sdk_version = sdk_version,
-            );
-        }
-
-        let duration_ms = (end - start).num_milliseconds() as f64;
-        if delay_ms >= 0 && duration_ms >= 0.0 {
-            let percentage = ((delay_ms as f64) / duration_ms).min(1.0) * 100.0;
-            metric!(
-                histogram(RelayHistograms::DynamicSamplingChangePercentage) = percentage,
-                source = source,
-                platform = platform,
-                sdk_name = sdk_name,
-                sdk_version = sdk_version,
-            );
-        }
     }
 }
 
@@ -1074,9 +962,16 @@ impl EnvelopeProcessorService {
             match relay_profiling::expand_profile(&item.payload()[..]) {
                 Ok(payloads) => new_profiles.extend(payloads),
                 Err(err) => {
-                    relay_log::debug!("invalid profile: {:#?}", err);
+                    match err {
+                        relay_profiling::ProfileError::InvalidJson(_) => {
+                            relay_log::warn!("invalid profile: {}", LogError(&err));
+                        }
+                        _ => relay_log::debug!("invalid profile: {}", err),
+                    };
                     context.track_outcome(
-                        outcome_from_profile_error(err),
+                        Outcome::Invalid(DiscardReason::Profiling(
+                            relay_profiling::discard_reason(err),
+                        )),
                         DataCategory::Profile,
                         1,
                     );
@@ -1135,6 +1030,11 @@ impl EnvelopeProcessorService {
                         }
                         Err(e) => {
                             relay_log::warn!("failed to parse replay event: {}", e);
+                            context.track_outcome(
+                                Outcome::Invalid(DiscardReason::InvalidReplayRecordingEvent),
+                                DataCategory::Replay,
+                                1,
+                            );
                         }
                     }
                 }
@@ -1737,7 +1637,7 @@ impl EnvelopeProcessorService {
             received_at: Some(envelope_context.received_at()),
             breakdowns: project_state.config.breakdowns_v2.clone(),
             span_attributes: project_state.config.span_attributes.clone(),
-            client_sample_rate: envelope.sampling_context().and_then(|ctx| ctx.sample_rate),
+            client_sample_rate: envelope.dsc().and_then(|ctx| ctx.sample_rate),
         };
 
         let mut store_processor = StoreProcessor::new(store_config, self.geoip_lookup.as_ref());
@@ -1750,6 +1650,42 @@ impl EnvelopeProcessorService {
         });
 
         Ok(())
+    }
+
+    /// Ensures there is a valid dynamic sampling context and corresponding project state.
+    ///
+    /// The dynamic sampling context (DSC) specifies the project_key of the project that initiated
+    /// the trace. That project state should have been loaded previously by the project cache and is
+    /// available on the `ProcessEnvelopeState`. Under these conditions, this cannot happen:
+    ///
+    ///  - There is no DSC in the envelope headers. This occurs with older or third-party SDKs.
+    ///  - The project key does not exist. This can happen if the project key was disabled, the
+    ///    project removed, or in rare cases when a project from another Sentry instance is referred
+    ///    to.
+    ///  - The project key refers to a project from another organization. In this case the project
+    ///    cache does not resolve the state and instead leaves it blank.
+    ///  - The project state could not be fetched. This is a runtime error, but in this case Relay
+    ///    should fall back to the next-best sampling rule set.
+    ///
+    /// In all of the above cases, this function will compute a new DSC using information from the
+    /// event payload, similar to how SDKs do this. The `sampling_project_state` is also switched to
+    /// the main project state.
+    ///
+    /// If there is no transaction event in the envelope, this function will do nothing.
+    fn normalize_dsc(&self, state: &mut ProcessEnvelopeState) {
+        if state.envelope.dsc().is_some() && state.sampling_project_state.is_some() {
+            return;
+        }
+
+        // The DSC can only be computed if there's a transaction event. Note that `from_transaction`
+        // below already checks for the event type.
+        let Some(event) = state.event.value() else { return };
+        let Some(key_config) = state.project_state.get_public_key_config() else { return };
+
+        if let Some(dsc) = DynamicSamplingContext::from_transaction(key_config.public_key, event) {
+            state.envelope.set_dsc(dsc);
+            state.sampling_project_state = Some(state.project_state.clone());
+        }
     }
 
     fn filter_event(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
@@ -1859,11 +1795,8 @@ impl EnvelopeProcessorService {
 
             state.transaction_metrics_extracted = true;
             state.envelope_context.set_event_metrics_extracted();
-
-            if let Some(context) = state.envelope.sampling_context() {
-                track_sampling_metrics(&state.project_state, context, event);
-            }
         }
+
         Ok(())
     }
 
@@ -1973,7 +1906,7 @@ impl EnvelopeProcessorService {
         let client_ip = state.envelope.meta().client_addr();
 
         match utils::should_keep_event(
-            state.envelope.sampling_context(),
+            state.envelope.dsc(),
             state.event.value(),
             client_ip,
             &state.project_state,
@@ -2053,6 +1986,7 @@ impl EnvelopeProcessorService {
 
             self.finalize_event(state)?;
             self.light_normalize_event(state)?;
+            self.normalize_dsc(state);
             self.filter_event(state)?;
             self.extract_transaction_metrics(state)?;
             self.sample_envelope(state)?;